requirements.txt
PyQt6
PyOpenGL
numpy
pandas
pyarrow